requirements.txt
netifaces
Pillow
psutil
pyTelegramBotAPI
requests
urllib3
winregistry
pypiwin32
tk
mathø
